pyproject.toml
[tool.poetry]
name = "dubdub"
version = "0.0.2"
description = ""
authors = ["Kevin Hill <kah.kevin.hill@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
colorama = "^0.4.4"
typer = { extras=["all"], version="^0.3.2" }
pydantic = "^1.8.2"
toolz = "^0.11.1"
plum-dispatch = "^1.3.1"
loguru = "^0.5.3"
ward = "^0.62.1-beta.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"
rich = "^10.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dubdub = 'dubdub.core:app'
